refactor: application 다음으로 검사하는 것을 최적화 하였습니다
- useSendAppliction -> useApplication
- canNext -> canApplicationNext(in useApplication)
- remove when no data in getApplicationValues
